proxy: separate handler and proxy
